refactor: extract ExtractedIssueSeverity.swift from IssueCoreSubtypes.swift
Splits the `ExtractedIssueSeverity` enum out of IssueCoreSubtypes.swift
so category + severity live in matching one-per-file layout.

Byte-identical move of lines 12-19. Both files import Foundation.

IssueCoreSubtypes.swift: 147 → 138 lines. Tests: 667.
