+++ Sources/BugNarrator/Models/IssueCoreSubtypes.swift
@@ -9,15 +9,6 @@ enum ExtractedIssueCategory: String, Codable, CaseIterable, Identifiable {
     var id: String { rawValue }
 }
 
-enum ExtractedIssueSeverity: String, Codable, CaseIterable, Identifiable {
-    case critical = "Critical"
-    case high = "High"
-    case medium = "Medium"
-    case low = "Low"
-
-    var id: String { rawValue }
-}
-
 struct IssueReproductionStep: Identifiable, Codable, Equatable {
     let id: UUID
     var instruction: String
